+++ src/components/functions/DevMenu.jsx
@@ -88,7 +88,7 @@ export function DevMenu() {
         <Dropdown.Menu onAction={handleAction}>
           {DEV_ACTIONS.map((action) => (
             <Dropdown.Item id={action.id} key={action.id} textValue={action.label}>
-              <action.icon className='size-5 shrink-0' stroke={1.5} />
+              <action.icon className='size-4 shrink-0' stroke={1.5} />
               <Label>{action.label}</Label>
             </Dropdown.Item>
           ))}